pyproject.toml
[tool.poetry]
name = "miaomiaonews"
version = "0.1.0"
description = "Help you follow a news topic with the latest development!"
authors = ["xw"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
newspaper3k = "^0.2.8"
pandas = "^2.0.0"
pydantic = {extras = ["email"], version = "^1.10.7"}
loguru = "^0.6.0"
matplotlib = "^3.7.1"
openai = "^0.27.4"
sqlalchemy= "^1.4.29"
fastapi = "^0.95.1"
pyjwt = "^2.6.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
uvicorn = "^0.21.1"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
notebook = "^6.5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
